Fix validation; unified VID type in dclauth - allow 0 for VID/PID/SoftwareVersion - use int32 instead of uint64 for VID in dclauth

+++ x/dclauth/handler_test.go
@@ -623,7 +623,7 @@ func storeTrustee(setup TestSetup) sdk.AccAddress {
 	return storeAccountWithVendorID(setup, types.Trustee, 0)
 }
 
-func storeAccountWithVendorID(setup TestSetup, role types.AccountRole, vendorID uint64) sdk.AccAddress {
+func storeAccountWithVendorID(setup TestSetup, role types.AccountRole, vendorID int32) sdk.AccAddress {
 	_, pubKey, address := testdata.KeyTestPubAddr()
 	ba := authtypes.NewBaseAccount(address, pubKey, 0, 0)
 	account := types.NewAccount(ba, types.AccountRoles{role}, vendorID)
